Made uri format escape special characters before parsing
Ruby's URI.parse does not accept non-latin characters, and will raise an
error on any uri that is not completely in the latin alphabet. This is a
problem for example, in japanese uris. I believe that the correct way to
handle this in ruby is to escape the string before parsing it.