Switched on quirks mode for json gem
This makes it behave:

1. More like multi json and yajl, which can parse scalar values (not
just hashes or arrays)
2. More correctly, in terms of the latest json rfc (rfc7159)*

Incidentally both multi_json and rails turn on quirks mode by default.

But there are a few other reasons for doing this - before, if json schema
was passed a string, then parsing would fail and the string would be
passed into the validator as a string. But that string could contain an
integer or a float value. That's clearly not the correct behavior.