Used minitest::benchmark to profile validation performance
I've had a first pass at benchmarking the gem. This adds a few basic
tests, to ensure that #fully_validate has constant performance as the
size of the json being parsed increases.

I'm not sure it's very useful at the moment. It will fail if any of the
tests do not run in constant time (as the size of the json increases)
but this isn't a problem right now. It would be more instructive to see
the results of this over time, as the codebase changes.