Use latest json-schema draft in tests by default
Lots of core tests were explicitly using draft-03, which works slightly
differently to the latest (eg. draft-04). I've updated all
tests to not specify the schema version at all (so they default to using
the latest).To me this feels safer, as (unless explicitly specified) we
will always testing against the default json-schema draft.

The main change has been the way that the "required" property is
defined (moving from the definition of each "property"
to being defined as an array alongside the "properties" object).