Made sure clear_cache also clears the cache of parsed uris
There are two main caches in json-schema. One is for schemas, one is for
parsed uris. Previously, setting `:clear_cache` as a validation option
would clear the schema cache but not the uri cache. This has been
reported to cause a memory leak (see issue #329) as over time the uri
cache gets bigger and bigger and is never cleared.

This change makes sure that whenever the schema cache is cleared, the
uri cache is also cleared.

Incidentally this change is not thread safe, and multithreaded
applications that clear the cache may see  issues. But until we can
refactor to either disable caching or get rid of class variables,
I don't believe there's much we can (realistically) do about that. This
is still an improvement on what's gone before.