Load local copy of draft schemas
If you refer to a draft schema in a $ref (as the common test suite does)
json-schema will load it from json-schema.org even though the gem
contains a local copy.

This change patches JSON::Validator#load_ref_schema to check whether a
$ref refers to a draft schema before trying to load from the web, and if
so parsing and serving up the local copy.

I've also replaced the local copy of draft3 with the exact version found
online (only whitespace was different).