Refactored common test suite ignored tests into a yaml file
With draft6 coming we have a lot more ignored tests (for now, until we
can fix the code to support new features). Rather than just add these
all onto the current hash of ignored tests, I've refactored them out
to a yaml file.

I've also changed how we ignore tests that we don't
support. Previously any test that we didn't support simply didn't have
a test created for it, in the test suite. Instead I have changed it so
that only optional tests are handled in this way. Non-optional (ie
core) tests that we don't support are skipped instead, so they are
visible to anyone running the tests, and so that we don't forget that
we should implement those at a future date.