Made it possible to run the tests without internet access
When the default rake task is run (which is the standard way to run
the tests) we always update the common test suite before running the
tests. Unfortunately, if the computer running the tests has no
internet access, this update will fail and then the tests never run.

I've changed this so that if the update to the common test suite
fails, we display an error then carry on with the rest of the test
suite.