Removed update_common_tests as a dependency of the test rake task
For quite while we've made the test rake task update the common test
suite every time it's run, to stop the test suite from becoming
out-of-date. However, the downside to this is that many pull requests
end up failing due to new tests in the common test suite, rather than
changes in the pull request itself.

I think it'd be best to manually update the common test suite in
future. It should make the test suite more reliable.