Corrected the draf6 schema id
Draft6 is a work in progress. The maintainers for now are using the
schema id "http://json-schema.org/draft/schema#". Previously I had
assumed it would follow the naming convention of the previous
drafts (and include the version number in the name). Unfortunately I
was wrong, so now I've updated it to use the correct schema id.

I've also updated the meta schema to the latest version.