Made sure we require timeout before using it
If you run `JSON::Validator.validate(metaschema, "http://a/%%30%30")`
then you'll see the error "uninitialized constant Timeout". This is
happening because we reference `Timeout` in `Validator` without
requiring it.

Interestingly I can't reproduce this error in the test suite because
something else (probably webmock?) is already requiring timeout for
us, and this error does not occur.

Fixes #384