Data that looks like a URI but is invalid should be treated as a string
There's a bug where if you pass a string as data and it looks like a
URI, we try to parse it and load the data from that location. However,
if parsing fails, we should give up trying to load remote data and
treat the data as a string instead.