Always download the common test suite before running tests
In #377 I made it possible to run the tests without having internet
access. But it seems that in the process I also disabled the code that
downloads the common test suite when you first check out the
codebase.

This changes the behaviour so that if you run the test suite and you
have internet access the common test suite is automatically
downloaded. If you don't have internet access it continues without
it. To update the common test suite to the latest version, you now
need to run the tests with the `UPDATE_TEST_SUITE` env var set.