Add failing test case for ref usage
This adds a failing test for a test case extracted from a local issue
with a JSON schema. This schema seems to work fine in other
implementations, such as https://www.jsonschemavalidator.net which says
it's valid and https://github.com/xeipuuv/gojsonschema which we also use
for the same original schema (I simplified here for reproduction
purposes).

I've tried further debugging why this doesn't seem to work, but have
failed so far. It seems though that a failing test case is still then a
useful bugreport this way.

Also happy for further dig in myself if someone has some pointers here
as to where to look and what could be wrong here.