Add caching to JSON::Util::URI.unescaped_path method
I decided not to re-raise the `Addressable::URI::InvalidURIError`, which was originally re-raised
as `JSON::Schema::UriError`. This decision was made because the subsequent call
to `Addressable::URI.unescape(uri)` does not handle these errors, requiring the user to handle
both types of errors.

Given that any call to the Addressable gem could potentially lead to an error, and
the JSON Schema gem does not consistently wrap them as `JSON::Schema::UriError`,

I decided to simplify the code for now. Error handling can be added later if necessary.